frontend/requirements.txt
# Core web framework
streamlit>=1.32.0
extra-streamlit-components>=0.1.81
# HTTP and web requests
requests>=2.31.0
beautifulsoup4>=4.12.3
fastapi>=0.104.0
uvicorn>=0.24.0
# Data processing and analysis
pandas>=2.0.0
numpy>=2.0.0
scipy>=1.11.0
# Visualization and plotting
plotly>=5.0.0
# Medical imaging
nibabel>=5.1.0
nilearn>=0.10.0
scikit-image>=0.21.0
dcm2niix>=1.0.0
# Progress bars and utilities
tqdm>=4.66.0
# Template rendering
jinja2>=3.1.0
# Environment and configuration
python-dotenv>=1.0.0
